+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-09-10T14:06:21.091Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-10T14:47:45.818Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")